refactor(core): move getFeaturedProducts closer to use
